fix(update): satisfy Windows batch lint boundaries
