+++ src/agents/apply-patch.test.ts
@@ -12,8 +12,7 @@ import {
   withRealpathSymlinkRebindRace,
 } from "../test-utils/symlink-rebind-race.js";
 import { createApplyPatchTool } from "./apply-patch.js";
-import { applyPatch } from "./apply-patch.test-support.js";
-import type { SandboxFsBridge } from "./sandbox/fs-bridge.js";
+import { applyPatch, createMemoryPatchSandbox } from "./apply-patch.test-support.js";
 
 async function withTempDir<T>(fn: (dir: string) => Promise<T>) {
   // realpath: production sandbox checks compare against canonical paths; on macOS
@@ -40,71 +39,6 @@ function buildAddFilePatch(targetPath: string): string {
 *** Add File: ${targetPath}
 +escaped
 *** End Patch`;
-}
-
-function createMemoryPatchSandbox(
-  initialFiles: Record<string, string | Buffer> = {},
-  options: { supportsExclusiveCreate?: boolean } = {},
-) {
-  const files = new Map<string, string | Buffer>(
-    Object.entries(initialFiles).map(([filePath, contents]) => [`/sandbox/${filePath}`, contents]),
-  );
-  const writeFile = vi.fn(async ({ filePath, data }) => {
-    files.set(filePath, Buffer.isBuffer(data) ? Buffer.from(data) : data);
-  });
-  const createFileExclusive = vi.fn(async ({ filePath, data }) => {
-    if (files.has(filePath)) {
-      return "exists" as const;
-    }
-    files.set(filePath, Buffer.isBuffer(data) ? Buffer.from(data) : data);
-    return "created" as const;
-  });
-  const mkdirp = vi.fn(async () => {});
-  const bridge: SandboxFsBridge = {
-    resolvePath: ({ filePath }) => ({
-      relativePath: filePath,
-      containerPath: `/sandbox/${filePath}`,
-    }),
-    readFile: async ({ filePath }) => {
-      const contents = files.get(filePath);
-      return typeof contents === "string"
-        ? Buffer.from(contents, "utf8")
-        : Buffer.from(contents ?? "");
-    },
-    writeFile,
-    ...(options.supportsExclusiveCreate === false ? {} : { createFileExclusive }),
-    remove: async ({ filePath }) => {
-      files.delete(filePath);
-    },
-    rename: async ({ from, to }) => {
-      const contents = files.get(from);
-      if (contents !== undefined) {
-        files.set(to, contents);
-        files.delete(from);
-      }
-    },
-    stat: async ({ filePath }) => {
-      const contents = files.get(filePath);
-      return contents === undefined
-        ? null
-        : { type: "file", size: Buffer.byteLength(contents), mtimeMs: 0 };
-    },
-    mkdirp,
-  };
-  return {
-    files,
-    bridge,
-    writeFile,
-    createFileExclusive,
-    mkdirp,
-    options: {
-      cwd: "/local/workspace",
-      sandbox: {
-        root: "/local/workspace",
-        bridge,
-      },
-    },
-  };
 }
 
 async function expectOutsideWriteRejected(params: {
@@ -201,15 +135,6 @@ describe("applyPatch", () => {
 
     expect(memory.files.get("/sandbox/hello.txt")).toBe("hello\n");
     expect(result.summary.added).toEqual(["hello.txt"]);
-  });
-
-  it("writes new Windows batch files with CRLF line endings", async () => {
-    const memory = createMemoryPatchSandbox();
-    await applyPatch(
-      "*** Begin Patch\n*** Add File: launch.cmd\n+@echo off\n+echo ready\n*** End Patch",
-      memory.options,
-    );
-    expect(memory.files.get("/sandbox/launch.cmd")).toBe("@echo off\r\necho ready\r\n");
   });
 
   it("rejects an add hunk that targets an existing file", async () => {
@@ -422,24 +347,6 @@ describe("applyPatch", () => {
     expect(memory.files.get("/sandbox/dest.txt")).toBe("foo\nbaz\n");
     expect(memory.files.has("/sandbox/source.txt")).toBe(false);
     expect(result.summary.modified).toEqual(["dest.txt"]);
-  });
-
-  it("normalizes moved Windows batch files to CRLF", async () => {
-    const memory = createMemoryPatchSandbox({
-      "source.txt": "@echo off\necho hello\n",
-    });
-    const patch = `*** Begin Patch
-*** Update File: source.txt
-*** Move to: launch.cmd
-@@
- @echo off
- echo hello
-*** End Patch`;
-
-    await applyPatch(patch, memory.options);
-
-    expect(memory.files.get("/sandbox/launch.cmd")).toBe("@echo off\r\necho hello\r\n");
-    expect(memory.files.has("/sandbox/source.txt")).toBe(false);
   });
 
   it("updates in place when move target resolves to the source file", async () => {
